feat(sandbox): host the last sample in the engine's loop
`Sandbox` is a struct with no fields. It has no simulation, no HUD and
no score, and it still runs, pauses, opens a menu, goes fullscreen and
reports a summary — which is the honest measure of how much of a frame
the engine now owns. Its `MenuAction` is `Infallible`, so
`MenuAction::Game` is uninhabited and the type system agrees that its
three buttons are the loop's.

It also stops declaring the six reserved keys for itself. They were
`crcbl::engine`'s constants already, and a second declaration is how
"the same key does the same thing in every sample" quietly stops being
true.

`FrameInfo::tick_dt` and `HostedGame::tick` widen to `f64`, which is
what `FrameClock::tick_dt_secs` reports. The sandbox is the only game
that reads it, and narrowing it was the engine picking a precision on
a game's behalf.

All five samples are converted; `apps/bare` stays hand-rolled and
`library_seam.rs` guards it. The backlog entry that has been tracking
this extraction since the pump slice is deleted — everything in it has
shipped.

+++ apps/asteroids/src/app.rs
@@ -256,7 +256,7 @@ impl HostedGame for Asteroids {
         menu::menus()
     }
 
-    fn tick(&mut self, _gpu: &mut Gpu, _tick_dt: f32) {
+    fn tick(&mut self, _gpu: &mut Gpu, _tick_dt: f64) {
         self.game.tick();
     }
