+++ apps/breakout/src/app.rs
@@ -232,7 +232,7 @@ impl HostedGame for Breakout {
         menu::menus()
     }
 
-    fn tick(&mut self, _gpu: &mut Gpu, _tick_dt: f32) {
+    fn tick(&mut self, _gpu: &mut Gpu, _tick_dt: f64) {
         self.game.tick();
     }
